module04/pyproject.toml
[project]
name = "module04"
version = "0.1.0"
description = "Analytics Engineering with dbt"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dbt-duckdb>=1.10.0",
]
